move chart containter to its own file
